test(admin-config): Move integration tests to "routes" package
This is a fixup for 33f711b.
