+++ components/admin-config/backend/src/test/kotlin/routes/InsertOrUpdateConfigIntegrationTest.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.eclipse.apoapsis.ortserver.components.adminconfig.endpoints
+package org.eclipse.apoapsis.ortserver.components.adminconfig.routes
 
 import io.kotest.assertions.ktor.client.shouldHaveStatus
 import io.kotest.matchers.shouldBe
@@ -31,13 +31,11 @@ import io.ktor.server.auth.authenticate
 import io.ktor.server.routing.routing
 
 import org.eclipse.apoapsis.ortserver.components.adminconfig.Config
-import org.eclipse.apoapsis.ortserver.components.adminconfig.routes.getConfigByKey
-import org.eclipse.apoapsis.ortserver.components.adminconfig.routes.setConfigByKey
 import org.eclipse.apoapsis.ortserver.shared.ktorutils.AbstractIntegrationTest
 
-class GetConfigByKeyIntegrationTest : AbstractIntegrationTest({
-    "GetConfigByKey" should {
-        "return the value and enabled status of an existing config key" {
+class InsertOrUpdateConfigIntegrationTest : AbstractIntegrationTest({
+    "InsertOrUpdateConfig" should {
+        "insert the new config key if it doesn't exist" {
             integrationTestApplication { client ->
                 application {
                     routing {
@@ -48,11 +46,10 @@ class GetConfigByKeyIntegrationTest : AbstractIntegrationTest({
                     }
                 }
 
-                // Insert a test config key to database
                 client.post("/admin/config/HOME_ICON_URL") {
                     setBody(
                         Config(
-                            value = "https://example.com/existing_icon.png",
+                            value = "https://example.com/icon.png",
                             isEnabled = true
                         )
                     )
@@ -62,20 +59,30 @@ class GetConfigByKeyIntegrationTest : AbstractIntegrationTest({
                 response shouldHaveStatus HttpStatusCode.OK
 
                 with(response.body<Config>()) {
-                    value shouldBe "https://example.com/existing_icon.png"
+                    value shouldBe "https://example.com/icon.png"
                     isEnabled shouldBe true
                 }
             }
         }
 
-        "return the default value if the config key does not exist in db" {
+        "update the value and isEnabled status of an existing config key" {
             integrationTestApplication { client ->
                 application {
                     routing {
                         authenticate("test") {
+                            setConfigByKey(dbExtension.db)
                             getConfigByKey(dbExtension.db)
                         }
                     }
+                }
+
+                client.post("/admin/config/HOME_ICON_URL") {
+                    setBody(
+                        Config(
+                            value = "https://example.com/icon.png",
+                            isEnabled = true
+                        )
+                    )
                 }
 
                 val response = client.get("/admin/config/HOME_ICON_URL")
@@ -83,7 +90,24 @@ class GetConfigByKeyIntegrationTest : AbstractIntegrationTest({
 
                 with(response.body<Config>()) {
                     value shouldBe "https://example.com/icon.png"
-                    isEnabled shouldBe false
+                    isEnabled shouldBe true
+                }
+
+                client.post("/admin/config/HOME_ICON_URL") {
+                    setBody(
+                        Config(
+                            value = "https://changed/example.com/explicit_icon.png",
+                            isEnabled = true
+                        )
+                    )
+                }
+
+                val response2 = client.get("/admin/config/HOME_ICON_URL")
+                response2 shouldHaveStatus HttpStatusCode.OK
+
+                with(response2.body<Config>()) {
+                    value shouldBe "https://changed/example.com/explicit_icon.png"
+                    isEnabled shouldBe true
                 }
             }
         }
